[Tests]: Fix and split out the tests into differents files.
